% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_upload.R
\name{data_upload_ui}
\alias{data_upload_ui}
\title{Data upload module UI}
\usage{
data_upload_ui(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\value{
A \code{tagList} with a file input and an upload button.
}
\description{
Data upload module UI
}
