% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preview_validation.R
\name{preview_validation_ui}
\alias{preview_validation_ui}
\title{Preview & validation module UI}
\usage{
preview_validation_ui(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\value{
A \code{tagList} with a data preview table and validation summary
boxes.
}
\description{
Preview & validation module UI
}
