% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visual_analytics.R
\name{visual_analytics_ui}
\alias{visual_analytics_ui}
\title{Visual analytics module UI}
\usage{
visual_analytics_ui(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\value{
A \code{tagList} with a date range input and an interactive expenses
plot.
}
\description{
Visual analytics module UI
}
